Sentiment Analysis on Twitter Data
The goal of this project is to develop a Twitter sentiment analysis model that can classify a given tweet as either "racist" or "non-racist." By utilizing the Naive Bayes algorithm, this project aims to analyze and categorize tweets based on their content, providing insight into whether the text exhibits racist language or sentiments. The model will be trained and tested on a dataset of tweets labeled with their corresponding categories, with the objective of achieving high accuracy in distinguishing between racist and non-racist content.
Comparison using the Random Forest and Support Vector Machine Algorithm used.